requirements.txt
biopython==1.78
matplotlib==3.3.3
numpy==1.19.5
pandas==1.1.5
pyparsing==2.4.7
python-dateutil==2.8.1
